Centralize run spend ledger accounting
